Month_report

+++ db/migrate/20190717023453_create_savings.rb
@@ -1,7 +1,7 @@
 class CreateSavings < ActiveRecord::Migration[5.2]
   def change
     create_table :savings do |t|
-      t.integer :savingInYear
+      t.float :savingInYear
 
       t.timestamps
     end
